BRRR {BRRR}	R Documentation

Easily Play Notification Sounds on any Platform

Description

This package contains one function, skrrrahh(), with one purpose: To make it easy to play notification sounds on whatever platform you are on. It is intended to be useful, for example, if you are running a long analysis in the background and want to know when it is ready.

Details

The package just contains one function skrrrahh, check it out to see what it does. For sound on Windows and MacOS BRRR depends on the audio package. For sound on Linux BRRR depends on that either the paplay utility from the Pulse Audio system, the aplay utility from the ALSA system, or VLC media player (http://www.videolan.org/vlc/index.html) is installed and on the PATH. Chances are that you alread have one of these.

Author(s)

Brooke Watson <brooke.meredith.watson@gmail.com>

Examples

# Play gucci saying "BRRR"
skrrrahh()

[Package BRRR version 1.2 Index]