skrrrahh {BRRR}	R Documentation

Play a short rap ad lib.

Description

skrrrahh is exactly like beep from the beepr package (https://cran.r-project.org/web/packages/beepr/index.html). The only difference is in the automatically loaded sounds available. skrrrahh uses rap adlibs from [The Rap Board](http://therapboard.com/). skrrrahh_list exposes a numbered listing of available sounds.

Usage

skrrrahh(sound = 27, expr = NULL)

skrrrahh_list()

Arguments

sound
character string or number specifying what sound to be played by either specifying one of the built in sounds, specifying the path to a wav file or specifying an url. The default is Gucci Mane saying "BRRR", sound 26. If sound does not match any of the loaded sounds, or is a valid path or url, a random sound will be played. Currently skrrrahh can only handle http urls, https is not supported.

expr
An optional expression to be executed before the sound.

Details

If skrrrahh is not able to play the sound, a warning is issued rather than an error. This is in order to not risk aborting or stopping the process that you wanted to get notified about.

Examples

# Play Gucci saying "BRRR"
skrrrahh()

# See the names of available sounds, along with their associated number
skrrrahh_list()

## Not run: 
# Get focused with Jay-Z.
skrrrahh(30)
# or
skrrrahh("jayz1")

# Play a random rap adlib.
skrrrahh(0)

# Update all packages and have Big Sean yell Whoa Dere when it is ready 
update.packages(ask=FALSE); skrrrahh(10)

# Change your options to have DJ Khaled yell: They dont wanna see us win!
# everytime you hit an error message.

options(error = function() {skrrrahh(34)})

## End(Not run)

[Package BRRR version 1.2 Index]